work on lesson 17

+++ Cartesian Plane Part 17/src/main/java/com/acmemail/judah/cartesian_plane/sandbox/jtable/panels/StateParser.java
@@ -83,7 +83,7 @@ public class StateParser
                 .peek( sa -> temp.setAbbreviation( sa[0] ) )
                 .peek( sa -> temp.setName( sa[1] ) )
                 .map( sa -> sa[2] )
-                .mapToDouble( Double::parseDouble )
+                .mapToInt( Integer::parseInt )
                 .peek( temp::setPopulation )
                 .mapToObj( d -> State.findByName( states, temp.getName() ) )
                 .filter( st -> st != null )
